Make DMA and OpenAL backends into proper APIs.
Reduces number of #ifdefs. Also isolates all DMA code in one place.
