+++ src/CMakeLists.txt
@@ -118,7 +118,6 @@ SET(SRC_CLIENT
 	client/sound/al.c
 	client/sound/main.c
 	client/sound/mem.c
-	client/sound/mix.c
 	client/sound/ogg.c
 	client/sound/qal/fixed.c
 #	client/sound/qal/dynamic.c
