+++ src/client/sound/main.c
@@ -29,8 +29,9 @@ int         s_registration_sequence;
 channel_t   s_channels[MAX_CHANNELS];
 int         s_numchannels;
 
-sndstarted_t s_started;
-bool        s_active;
+sndstarted_t    s_started;
+bool            s_active;
+sndapi_t        s_api;
 
 vec3_t      listener_origin;
 vec3_t      listener_forward;
@@ -75,15 +76,7 @@ static void S_SoundInfo_f(void)
         return;
     }
 
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        AL_SoundInfo();
-#endif
-
-#if USE_SNDDMA
-    if (s_started == SS_DMA)
-        DMA_SoundInfo();
-#endif
+    s_api.sound_info();
 }
 
 static void S_SoundList_f(void)
@@ -159,13 +152,17 @@ void S_Init(void)
     s_started = SS_NOT;
 
 #if USE_OPENAL
-    if (s_started == SS_NOT && s_enable->integer >= SS_OAL && AL_Init())
+    if (s_started == SS_NOT && s_enable->integer >= SS_OAL && snd_openal.init()) {
         s_started = SS_OAL;
+        s_api = snd_openal;
+    }
 #endif
 
 #if USE_SNDDMA
-    if (s_started == SS_NOT && s_enable->integer >= SS_DMA && DMA_Init())
+    if (s_started == SS_NOT && s_enable->integer >= SS_DMA && snd_dma.init()) {
         s_started = SS_DMA;
+        s_api = snd_dma;
+    }
 #endif
 
     if (s_started == SS_NOT) {
@@ -204,10 +201,8 @@ fail:
 
 static void S_FreeSound(sfx_t *sfx)
 {
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        AL_DeleteSfx(sfx);
-#endif
+    if (s_api.delete_sfx)
+        s_api.delete_sfx(sfx);
     if (sfx->cache)
         Z_Free(sfx->cache);
     if (sfx->truename)
@@ -241,15 +236,8 @@ void S_Shutdown(void)
 	OGG_SaveState();
 	OGG_Shutdown();
 
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        AL_Shutdown();
-#endif
-
-#if USE_SNDDMA
-    if (s_started == SS_DMA)
-        DMA_Shutdown();
-#endif
+    s_api.shutdown();
+    memset(&s_api, 0, sizeof(s_api));
 
     s_started = SS_NOT;
     s_active = false;
@@ -279,18 +267,8 @@ void S_Activate(void)
     s_active = active;
 
     Com_DDDPrintf("%s: %d\n", __func__, s_active);
-
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        S_StopAllSounds();
-#endif
-
-#if USE_SNDDMA
-    if (s_started == SS_DMA)
-        DMA_Activate();
-#endif
+    s_api.activate();
 }
-
 
 // =======================================================================
 // Load a sound
@@ -506,9 +484,6 @@ void S_EndRegistration(void)
 {
     int     i;
     sfx_t   *sfx;
-#if USE_SNDDMA
-    sfxcache_t *sc;
-#endif
 
     S_RegisterSexedSounds();
 
@@ -524,16 +499,9 @@ void S_EndRegistration(void)
             S_FreeSound(sfx);
             continue;
         }
-#if USE_OPENAL
-        if (s_started == SS_OAL)
-            continue;
-#endif
-#if USE_SNDDMA
         // make sure it is paged in
-        sc = sfx->cache;
-        if (sc)
-            Com_PageInMemory(sc->data, sc->size);
-#endif
+        if (s_api.page_in_sfx)
+            s_api.page_in_sfx(sfx);
     }
 
     // load everything in
@@ -595,94 +563,12 @@ channel_t *S_PickChannel(int entnum, int entchannel)
         return NULL;
 
     ch = &s_channels[first_to_die];
-#if USE_OPENAL
-    if (s_started == SS_OAL && ch->sfx)
-        AL_StopChannel(ch);
-#endif
+    if (s_api.stop_channel)
+        s_api.stop_channel(ch);
     memset(ch, 0, sizeof(*ch));
 
     return ch;
 }
-
-#if USE_SNDDMA
-
-/*
-=================
-S_SpatializeOrigin
-
-Used for spatializing channels and autosounds
-=================
-*/
-void S_SpatializeOrigin(const vec3_t origin, float master_vol, float dist_mult, int *left_vol, int *right_vol)
-{
-    vec_t       dot;
-    vec_t       dist;
-    vec_t       lscale, rscale, scale;
-    vec3_t      source_vec;
-
-    if (cls.state != ca_active) {
-        *left_vol = *right_vol = 255;
-        return;
-    }
-
-// calculate stereo seperation and distance attenuation
-    VectorSubtract(origin, listener_origin, source_vec);
-
-    dist = VectorNormalize(source_vec);
-    dist -= SOUND_FULLVOLUME;
-    if (dist < 0)
-        dist = 0;           // close enough to be at full volume
-    dist *= dist_mult;      // different attenuation levels
-
-    dot = DotProduct(listener_right, source_vec);
-
-    if (dma.channels == 1 || !dist_mult) {
-        // no attenuation = no spatialization
-        rscale = 1.0f;
-        lscale = 1.0f;
-    } else {
-        rscale = 0.5f * (1.0f + dot);
-        lscale = 0.5f * (1.0f - dot);
-    }
-
-    // add in distance effect
-    scale = (1.0f - dist) * rscale;
-    *right_vol = (int)(master_vol * scale);
-    if (*right_vol < 0)
-        *right_vol = 0;
-
-    scale = (1.0f - dist) * lscale;
-    *left_vol = (int)(master_vol * scale);
-    if (*left_vol < 0)
-        *left_vol = 0;
-}
-
-/*
-=================
-S_Spatialize
-=================
-*/
-static void S_Spatialize(channel_t *ch)
-{
-    vec3_t      origin;
-
-    // anything coming from the view entity will always be full volume
-    if (ch->entnum == -1 || ch->entnum == listener_entnum) {
-        ch->leftvol = ch->master_vol * 255;
-        ch->rightvol = ch->master_vol * 255;
-        return;
-    }
-
-    if (ch->fixed_origin) {
-        VectorCopy(ch->origin, origin);
-    } else {
-        CL_GetEntitySoundOrigin(ch->entnum, origin);
-    }
-
-    S_SpatializeOrigin(origin, ch->master_vol * 255, ch->dist_mult, &ch->leftvol, &ch->rightvol);
-}
-
-#endif
 
 /*
 =================
@@ -766,19 +652,10 @@ void S_IssuePlaysound(playsound_t *ps)
     ch->sfx = ps->sfx;
     VectorCopy(ps->origin, ch->origin);
     ch->fixed_origin = ps->fixed_origin;
-
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        AL_PlayChannel(ch);
-#endif
-
-#if USE_SNDDMA
-    if (s_started == SS_DMA)
-        S_Spatialize(ch);
-#endif
-
     ch->pos = 0;
     ch->end = s_paintedtime + sc->length;
+
+    s_api.play_channel(ch);
 
     // free the playsound
     S_FreePlaysound(ps);
@@ -841,16 +718,7 @@ void S_StartSound(const vec3_t origin, int entnum, int entchannel, qhandle_t hSf
     ps->attenuation = attenuation;
     ps->volume = vol;
     ps->sfx = sfx;
-
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        ps->begin = s_paintedtime + timeofs * 1000;
-#endif
-
-#if USE_SNDDMA
-    if (s_started == SS_DMA)
-        ps->begin = DMA_DriftBeginofs(timeofs);
-#endif
+    ps->begin = s_api.set_begin_ofs(timeofs);
 
     // sort into the pending sound list
     for (sort = s_pendingplays.next; sort != &s_pendingplays && sort->begin < ps->begin; sort = sort->next)
@@ -925,15 +793,7 @@ void S_StopAllSounds(void)
         s_playsounds[i].next->prev = &s_playsounds[i];
     }
 
-#if USE_OPENAL
-    if (s_started == SS_OAL)
-        AL_StopAllChannels();
-#endif
-
-#if USE_SNDDMA
-    if (s_started == SS_DMA)
-        DMA_ClearBuffer();
-#endif
+    s_api.stop_all_sounds();
 
     // clear all the channels
     memset(s_channels, 0, sizeof(s_channels));
@@ -962,92 +822,6 @@ void S_BuildSoundList(int *sounds)
     }
 }
 
-#if USE_SNDDMA
-
-/*
-==================
-S_AddLoopSounds
-
-Entities with a ->sound field will generated looped sounds
-that are automatically started, stopped, and merged together
-as the entities are sent to the client
-==================
-*/
-static void S_AddLoopSounds(void)
-{
-    int         i, j;
-    int         sounds[MAX_EDICTS];
-    int         left, right, left_total, right_total;
-    channel_t   *ch;
-    sfx_t       *sfx;
-    sfxcache_t  *sc;
-    int         num;
-    entity_state_t  *ent;
-    vec3_t      origin;
-
-    if (cls.state != ca_active || !s_active || sv_paused->integer || !s_ambient->integer) {
-        return;
-    }
-
-    S_BuildSoundList(sounds);
-
-    for (i = 0; i < cl.frame.numEntities; i++) {
-        if (!sounds[i])
-            continue;
-
-        sfx = S_SfxForHandle(cl.sound_precache[sounds[i]]);
-        if (!sfx)
-            continue;       // bad sound effect
-        sc = sfx->cache;
-        if (!sc)
-            continue;
-
-        num = (cl.frame.firstEntity + i) & PARSE_ENTITIES_MASK;
-        ent = &cl.entityStates[num];
-
-        // find the total contribution of all sounds of this type
-        CL_GetEntitySoundOrigin(ent->number, origin);
-        S_SpatializeOrigin(origin, 255.0f, SOUND_LOOPATTENUATE,
-                           &left_total, &right_total);
-        for (j = i + 1; j < cl.frame.numEntities; j++) {
-            if (sounds[j] != sounds[i])
-                continue;
-            sounds[j] = 0;  // don't check this again later
-
-            num = (cl.frame.firstEntity + j) & PARSE_ENTITIES_MASK;
-            ent = &cl.entityStates[num];
-
-            CL_GetEntitySoundOrigin(ent->number, origin);
-            S_SpatializeOrigin(origin, 255.0f, SOUND_LOOPATTENUATE,
-                               &left, &right);
-            left_total += left;
-            right_total += right;
-        }
-
-        if (left_total == 0 && right_total == 0)
-            continue;       // not audible
-
-        // allocate a channel
-        ch = S_PickChannel(0, 0);
-        if (!ch)
-            return;
-
-        if (left_total > 255)
-            left_total = 255;
-        if (right_total > 255)
-            right_total = 255;
-        ch->leftvol = left_total;
-        ch->rightvol = right_total;
-        ch->master_vol = 1.0f;
-        ch->autosound = true;   // remove next frame
-        ch->sfx = sfx;
-        ch->pos = s_paintedtime % sc->length;
-        ch->end = s_paintedtime + sc->length - ch->pos;
-    }
-}
-
-#endif
-
 /*
 ============
 S_Update
@@ -1057,11 +831,6 @@ Called once each time through the main loop
 */
 void S_Update(void)
 {
-#if USE_SNDDMA
-    int         i;
-    channel_t   *ch;
-#endif
-
     if (cvar_modified & CVAR_SOUND) {
         Cbuf_AddText(&cmd_buffer, "snd_restart\n");
         cvar_modified &= ~CVAR_SOUND;
@@ -1087,61 +856,8 @@ void S_Update(void)
         listener_entnum = cl.frame.clientNum + 1;
     }
 
-#if USE_OPENAL
-    if (s_started == SS_OAL) {
-		OGG_Stream();
-		AL_Update();
-        return;
-    }
-#endif
-
-#if USE_SNDDMA
-    // rebuild scale tables if volume is modified
-    if (s_volume->modified)
-        S_InitScaletable();
-
-    // update spatialization for dynamic sounds
-    for (i = 0, ch = s_channels; i < s_numchannels; i++, ch++) {
-        if (!ch->sfx)
-            continue;
-        if (ch->autosound) {
-            // autosounds are regenerated fresh each frame
-            memset(ch, 0, sizeof(*ch));
-            continue;
-        }
-        S_Spatialize(ch);         // respatialize channel
-        if (!ch->leftvol && !ch->rightvol) {
-            memset(ch, 0, sizeof(*ch));
-            continue;
-        }
-    }
-
-    // add loopsounds
-    S_AddLoopSounds();
-
     OGG_Stream();
-
-#ifdef USE_DEBUG
-    //
-    // debugging output
-    //
-    if (s_show->integer) {
-        int total = 0;
-        for (i = 0, ch = s_channels; i < s_numchannels; i++, ch++)
-            if (ch->sfx && (ch->leftvol || ch->rightvol)) {
-                Com_Printf("%3i %3i %s\n", ch->leftvol, ch->rightvol, ch->sfx->name);
-                total++;
-            }
-
-        if (s_show->integer > 1 || total) {
-            Com_Printf("----(%i)---- painted: %i\n", total, s_paintedtime);
-        }
-    }
-#endif
-
-// mix some sound
-    DMA_Update();
-#endif
+    s_api.update();
 }
 
 float S_GetLinearVolume(float perceptual)
@@ -1162,3 +878,19 @@ float S_GetLinearVolume(float perceptual)
     return volume;
 }
 
+/*
+ * Cinematic streaming and voice over network.
+ * This could be used for chat over network, but
+ * that would be terrible slow.
+ */
+void
+S_RawSamples(int samples, int rate, int width,
+    int channels, byte *data, float volume)
+{
+    s_api.raw_samples(samples, rate, width, channels, data, volume);
+}
+
+void S_UnqueueRawSamples()
+{
+    s_api.unqueue_raw_samples();
+}
